Update PlainRfcPreparedFunction.cs
The function.ToString() within the catch statement will raise an object ref not set" error because in the event that the RFC function couldn't be retrieved by the SAP RFC library then function == null. Add the safety check so that the correct error will be cascaded down to error catching at upper levels.